+++ backtrace-library/src/main/java/backtraceio/library/interfaces/Client.java
@@ -12,4 +12,9 @@ public interface Client {
      * @param report data which should be send to Backtrace API
      */
     void send(BacktraceReport report);
+
+    /**
+     * Capture unhandled native exceptions (Backtrace database integration is required to enable this feature).
+     */
+    void enableNativeIntegration();
 }